requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.3
dj3-cloudinary-storage==0.0.6
Django==4.0.3
django-allauth==0.49.0
django-anymail==8.5
django-cors-headers==3.11.0
django-countries==7.3.1
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
redis==4.2.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.0
xlwt==1.3.0
